Refinement code & Update (#84)
* Refinement code. Now all we have to do is to put in the effects.

Bugs to be expected! (Not really, but you're all warned)

Things to put in/code: Staff upgrades, Deathly Dagger, Parthia, bunny weapons, armor effective weapons (just to put the special tag in) and all the special ones that aren't in refinements.json yet.

* Refinery update completed + fixes + new characters + blessings

I tried testing this as much as possible! (All the weapons' refinements and blessings have been tested)

* Fixed missing 's' in lyn-wind{s}-embrace